+++ src/gpu/ResourceCache.h
@@ -106,7 +106,7 @@ class ResourceCache {
   std::list<Resource*> nonpurgeableResources = {};
   std::list<Resource*> purgeableResources = {};
   BytesKeyMap<std::vector<Resource*>> recycleKeyMap = {};
-  std::unordered_map<uint32_t, Resource*> resourceKeyMap = {};
+  std::unordered_map<uint64_t, Resource*> resourceKeyMap = {};
 
   static void AddToList(std::list<Resource*>& list, Resource* resource);
   static void RemoveFromList(std::list<Resource*>& list, Resource* resource);
